Added options (moved notation inside player.options)
I created player.options, which contains player.options.notation and player.options.hotkeys. This just makes it easier and bunches together options into a single thing. However, if you don't want this, that's fine. Just make it go back, and make sure to change hotkeys into something, too. Hotkeys is a boolean, if true, hotkeys are enabled. If not, they're not.